Move policy matching logic to virt-controller
Also, MigrationConfiguration is now part of vmi.Status.MigrationState.
Now virt-handler simply checks whether MigrationState's MigrationConfigs
are nil. If they aren't - it uses them, if they are - it uses
clusterConfig's MigrationConfigs.

Also Adjust unit tests and move them to migration_test.go
