+++ api/openapi-spec/swagger.json
@@ -14719,8 +14719,12 @@
       "description": "Indicates that the migration failed",
       "type": "boolean"
      },
-     "migrationConfigSource": {
-      "description": "Lets us know if the vmi migration configuration is affected by KubevirtCR or a migration policy",
+     "migrationConfiguration": {
+      "description": "Migration configurations to apply",
+      "$ref": "#/definitions/v1.MigrationConfiguration"
+     },
+     "migrationPolicyName": {
+      "description": "Name of the migration policy. If string is empty, no policy is matched",
       "type": "string"
      },
      "migrationUid": {
